fix(loco-command): fix msg_id type

+++ crates/talk-loco-command/src/request/chat/write.rs
@@ -17,7 +17,7 @@ pub struct WriteReq {
     ///
     /// Client send count??
     #[serde(rename = "msgId")]
-    pub msg_id: i32,
+    pub msg_id: i64,
 
     /// Message content
     ///
